requirements.txt
asgiref
attrs
Django==3.2
djangorestframework
drf-spectacular
inflection
jsonschema
jsonschema-specifications
psycopg2
pytz
PyYAML
referencing
sqlparse
tzdata
uritemplate
djangorestframework-simplejwt